eslint config modification and continue to work on facebook auth
