+++ .eslintrc.js
@@ -3,9 +3,11 @@ module.exports = {
     'es6': true
   },
   parserOptions: {
+    sourceType: "module",
     ecmaFeatures: {
       'jsx': true
     }
+
   },
   rules: {
     /* Indentation */
